created setUpLayout function; returns HTMLDivElement which refers to page body while having aleady checked for user token, displaying navbar, and setting up page background
